Move UFS out of meteor packages
