+++ apps/meteor/.meteor/versions
@@ -51,9 +51,6 @@ htmljs@1.1.1
 http@2.0.0
 id-map@1.1.1
 inter-process-messaging@0.1.1
-jalik:ufs@1.0.4
-jalik:ufs-gridfs@1.0.5
-jalik:ufs-local@1.0.4
 jparker:crypto-core@0.1.0
 jparker:crypto-md5@0.1.1
 jparker:gravatar@0.5.1
