+++ apps/meteor/app/file-upload/server/config/GridFS.js
@@ -2,8 +2,7 @@ import stream from 'stream';
 import zlib from 'zlib';
 import util from 'util';
 
-import { UploadFS } from 'meteor/jalik:ufs';
-
+import { UploadFS } from '../../../../server/ufs';
 import { Logger } from '../../../logger/server';
 import { FileUploadClass, FileUpload } from '../lib/FileUpload';
 import { getFileRange, setRangeHeaders } from '../lib/ranges';
