+++ apps/meteor/app/file-upload/server/config/_configUploadStorage.js
@@ -1,6 +1,6 @@
-import { UploadFS } from 'meteor/jalik:ufs';
 import _ from 'underscore';
 
+import { UploadFS } from '../../../../server/ufs';
 import { settings } from '../../../settings/server';
 import { SystemLogger } from '../../../../server/lib/logger/system';
 import './AmazonS3.js';
