+++ apps/meteor/app/models/client/index.ts
@@ -1,7 +1,4 @@
 import { Base } from './models/Base';
-import Avatars from './models/Avatars';
-import Uploads from './models/Uploads';
-import UserDataFiles from './models/UserDataFiles';
 import { Roles } from './models/Roles';
 import { Users } from './models/Users';
 import { CachedChannelList } from './models/CachedChannelList';
@@ -44,9 +41,6 @@ Meteor.userAsync = async () => {
 
 export {
 	Base,
-	Avatars,
-	Uploads,
-	UserDataFiles,
 	Roles,
 	CachedChannelList,
 	CachedChatRoom,
