--- apps/meteor/app/models/client/models/Avatars.ts
@@ -1,10 +1,0 @@
-import { Base } from './Base';
-
-class Avatars extends Base {
-	constructor() {
-		super();
-		this._initModel('avatars');
-	}
-}
-
-export default new Avatars();
