--- apps/meteor/packages/meteor-jalik-ufs/.gitignore
@@ -1,2 +1,0 @@
-.npm
-.idea
